% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind factors}
\usage{
fbind(...)
}
\arguments{
\item{...}{Two or more \code{factor()}s you want to bind together.}
}
\value{
A \code{factor()}
}
\description{
Combine factors together by finding their common levels and re-making
the factor.
}
\examples{

## Create some data
a <- factor(letters[1:4])
b <- factor(letters[3:6])
x <- factor("leo", "leonardo")

## Use fbind() with 2 or more factors
fbind(a, b)
fbind(a, b, x)
}
